webServ: Fix spam bug: - send ack instead of relay status
Fixes a bug on app side where if you spam the buttons it will loop on/off nonstop.
No need to publish relay status as we have req_relay for that.

send ack,ack instead because if we don`t send anything app crashes due to arrayIndex something and I have no clue how to fix that